plugins/mkdocs-uktre-glossary-plugin/pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project.entry-points."mkdocs.plugins"]
"uktre-glossary" = "mkdocs_uktre_glossary_plugin.plugin:GlossaryPlugin"

[project]
name="mkdocs_uktre_glossary_plugin"
keywords = ["mkdocs", "plugin"]
authors = [
    {name = "UKTRE"},
]
license = { text = "MIT" }

description="MkDocs plugin to generate UKTRE glossary"
readme = { file = "README.md", content-type = "text/markdown" }

requires-python=">=3.11"

classifiers=[
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]

dynamic = ["version"]
dependencies = [
    "mkdocs>=1.6.1",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "tabulate>=0.9.0",
    # We're hooking in to the internals of this plugin, so pin exactly
    "mkdocs-table-reader-plugin==3.1.0",
]

[project.urls]
"Homepage" = "https://github.com/TODO/TODO"

[tool.setuptools.dynamic]
version = {attr = "mkdocs_uktre_glossary_plugin.__version__"}
